requirements.txt
psycopg2-binary~=2.9.9
python-dotenv~=1.0.1
uvicorn~=0.32.0
fastapi~=0.115.4
Jinja2==3.1.4
python-multipart==0.0.17